Reorganize reaching definitions analysis.
