+++ angr/analyses/reaching_definitions/constants.py
@@ -1,0 +1,6 @@
+
+#
+# Observation point types
+#
+OP_BEFORE = 0
+OP_AFTER = 1
