+++ angr/analyses/reaching_definitions/undefined.py
@@ -1,0 +1,52 @@
+
+class Undefined(object):
+    def __init__(self):
+        pass
+
+    def __add__(self, other):
+        return self
+
+    def __radd__(self, other):
+        return self
+
+    def __sub__(self, other):
+        return self
+
+    def __rsub__(self, other):
+        return self
+
+    def __lshift__(self, other):
+        return self
+
+    def __rlshift__(self, other):
+        return self
+
+    def __rshift__(self, other):
+        return self
+
+    def __rrshift__(self, other):
+        return self
+
+    def __and__(self, other):
+        return self
+
+    def __rand__(self, other):
+        return self
+
+    def __xor__(self, other):
+        return self
+
+    def __rxor__(self, other):
+        return self
+
+    def __or__(self, other):
+        return self
+
+    def __ror__(self, other):
+        return self
+
+    def __neg__(self):
+        return self
+
+    def __eq__(self, other):
+        return self is other
